+++ ci/release_notes.md
@@ -2,3 +2,5 @@
 
 - Plugins now accept boolish strings and numbers in place of
   actual booleans.
+
+- Handle symlinks in the `fs` plugin